+++ springboot-shiro/src/main/java/com/hongfei/springbootshiro/user/model/dto/OrganizationDto.java
@@ -14,5 +14,5 @@ public class OrganizationDto {
     private String fullName;
     private String description;
     private Long parentId;
-    private Byte status;
+    private int status;
 }
